ODID_OceanDetectionwithImprovedDeNoising/requirements.txt
pyyaml
tqdm
tensorboard
opencv-python
numpy
torchvision
# torch is environment-specific (CPU/CUDA build), install from the official PyTorch channel
